refactored the project structure
