+++ scraper/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the scraper directory a Python package
