+++ scraper/utils/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the utils directory a Python package
